Making y a double variable instead of a float one

+++ Floats_and_IOMANIP.cpp
@@ -8,7 +8,7 @@ int main(){
 	cout<<x<<endl;	//printing the float variable
 	
 	
-	float y=23.456;	//declaring a float variable with 3 decimal positions
+	double y=23.456;	//declaring a float variable with 3 decimal positions
 	cout<<y<<endl;	//printing the whole number stored in y
 	
 	cout<<setprecision(2)<<fixed<<y<<endl; //printing the variable y only with 2 decimal positions
